Explain how to use the chromium revision
My coworker and I weren't sure what the "chromium revision" actually meant, and googling for "chromium revision" mostly brings up Github issues about puppeteer itself. I searched the repo and found this:

> BrowserFetcher operates on revision strings that specify a precise version of Chromium, e.g. `"533271"`. Revision strings can be obtained from [omahaproxy.appspot.com](http://omahaproxy.appspot.com/).

Which was helpful, so incorporated that info into the README. If there's more info to share about the significance of the chromium revision it'd be great to share that too!